[checkpoints] Simplify checkpoint roots calculation
Because we used to have 'streaming' interface from narwhal to sui, we needed two tables (`consensus_message_order` and `checkpoint_boundary`) to derive checkpoint roots on the commit boundary.

However, now that the interface is simplified, and we have an entire commit, those tables are not needed. Bunch of code is removed and DB writes are reduced